Refactor the hook and context - hook is using the context now
