+++ src/components/menu-components/card/Card.jsx
@@ -2,11 +2,13 @@ import "./Card.css"
 import foodImage from "../../../assets/menu-image.png"
 import addButton from "../../../assets/add-button.svg"
 import convertPrice from "../../../helpers/convertPrice.js";
-import {useContext} from "react";
-import {OrderContext} from "../../../context/OrderProvider.jsx";
+import useFetchOrderItem
+  from "../../../custom-hooks/api-requests/GET/UseFetchOrderItem.jsx";
+
 function Card({className, name, image, description, price, id}) {
 
-  const {addMealToOrder} = useContext(OrderContext)
+  const {addMealToOrder} = useFetchOrderItem();
+
   return (
     <article
       className={`card${className ? className : ''}`}
@@ -25,8 +27,6 @@ function Card({className, name, image, description, price, id}) {
         </div>
         <div className="price">€{convertPrice(price)}</div>
       </figcaption>
-
-
     </article>
   )
 }
